Update hw7
- Изменение типа аттрибута с numeric на float в ddl скрипте
- Обновление view management

+++ docker/postgres/init/seeder.sql
@@ -9,7 +9,7 @@ INSERT INTO movies (name) VALUES
 INSERT INTO attribute_types (type) VALUES
 ('integer'),
 ('text'),
-('numeric'),
+('float'),
 ('timestamp'),
 ('boolean')
 ;
